Add group colors, tag badges, board UUIDs, event cancellation, and streamline member identity
Features:
- Group colors and tag badges with position-based priority
- MemberIdentity carries display metadata (nameColor, displayTag, stationName)
- MemberName component renders avatar, colored name, tag badge, and
  federation station badge from identity
- Board UUID for stable federation references, overrides keyed by UUID
- Federated board member assignment endpoint
- Event cancellation with notification to registered members
- Auto-cancel events below minimum registration threshold via scheduler
- Layered EmberLogo replaces old EmberLogo everywhere, with bounce and
  FAQ shake animations
- Pride month text gets black stroke for readability

Infrastructure:
- Caffeine cache replaces ConcurrentHashMap in resolvers
- MemberIdentityFactory enriches all identities with display metadata
- MemberWithName shared record for consistent member API responses
- Avatar endpoint uses stationUid/memberUid path for federation support
- Dev error files cleared on startup
- Station export/import extended with 45+ tables (boards, news, events,
  attendance, equipment, forms, KB, inventory, restrictions, settings)
- Federated board operations pass actor identity for proper history

+++ src/main/java/dev/chojo/ember/feature/system/repository/ProblemReportRepository.java
@@ -60,10 +60,10 @@ public class ProblemReportRepository {
     }
 
     public int acknowledgeAll() {
-        return (int) Query.query("UPDATE problem_report SET acknowledged = TRUE WHERE acknowledged = FALSE;")
-                .single(Call.of())
-                .update()
-                .rows();
+        return Query.query("UPDATE problem_report SET acknowledged = TRUE WHERE acknowledged = FALSE;")
+                    .single(Call.of())
+                    .update()
+                    .rows();
     }
 
     public boolean delete(int id) {
